arb-analytics-prefect/requirements.txt
prefect>=2.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
prefect_github
